refactor(1013): extract EndpointConfig dataclass (SC-001, position 16)
Move the 10-LOC ``EndpointConfig`` dataclass out of MistHelper.py into
``src/dataclasses/endpoint_config.py``. The class is a pure descriptor used
exclusively by ``ConstDefinitionsExporter`` to describe each discovered Mist
const endpoint.

MistHelper.py keeps a re-export shim so ``MistHelper.EndpointConfig`` continues
to resolve, and the in-file references in ``ConstDefinitionsExporter`` bind to
the same symbol via that re-export -- no callsite changes required.

Local gate: black + ruff clean, ``py_compile`` OK, vulture clean,
``python MistHelper.py --test`` = 66 successful / 0 failed / 131 skipped.
